Backend/requirements.txt
django==4.2.11
python-dotenv==1.0.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
faker==23.3.0
openai>=1.0.0
